+++ test/int-sender.py
@@ -6,19 +6,35 @@ import sys
 import time
 import traceback
 
-from nrf24 import *
 import pigpio
+from nrf24 import *
 
 #
 # A simple NRF24L sender that connects to a PIGPIO instance on a hostname and port, default "localhost" and 8888, and
-# starts sending data on the address specified with a fixed payload size of 9 bytes.  
-# Use the companion program "fixed-receiver.py" to receive the data from it on a different Raspberry Pi.
+# starts sending data on the address specified.  Use the companion program "int-receiver.py" or "simple-receiver.py" to 
+# receive the data from it on a different Raspberry Pi.
 #
+def gpio_interrupt(gpio, level, tick):
+
+    # Interrupt information.
+    print(f"Interrupt: gpio={gpio}, level={['LOW', 'HIGH', 'NONE'][level]}, tick={tick}")
+
+    # Check result of last send operation.
+    if nrf.get_packages_lost() == 0:
+        print(f"Success: lost={nrf.get_packages_lost()}, retries={nrf.get_retries()}")
+    else:
+        print(f"Error: lost={nrf.get_packages_lost()}, retries={nrf.get_retries()}")
+    
+    # Reset and enter RX mode.
+    nrf.reset_packages_lost()
+    nrf.power_up_rx()
+
+
 if __name__ == "__main__":    
-    print("Python NRF24 Fixed Sender Example.")
+    print("Python NRF24 Simple Interrupt Based Sender Example.")
     
     # Parse command line argument.
-    parser = argparse.ArgumentParser(prog="fixed-sender.py", description="Simple NRF24 transmitter with fixed payload.")
+    parser = argparse.ArgumentParser(prog="int-sender.py", description="NRF24 Interrupt Based Sender Example.")
     parser.add_argument('-n', '--hostname', type=str, default='localhost', help="Hostname for the Raspberry running the pigpio daemon.")
     parser.add_argument('-p', '--port', type=int, default=8888, help="Port number of the pigpio daemon.")
     parser.add_argument('address', type=str, nargs='?', default='1SNSR', help="Address to send to (3 to 5 ASCII characters).")
@@ -37,20 +53,23 @@ if __name__ == "__main__":
     pi = pigpio.pi(hostname, port)
     if not pi.connected:
         print("Not connected to Raspberry Pi ... goodbye.")
-        exit()
+        sys.exit()
+
+    # Setup callback for interrupt on falling edge.
+    pi.callback(24, pigpio.FALLING_EDGE, gpio_interrupt)
 
     # Create NRF24 object.
     # PLEASE NOTE: PA level is set to MIN, because test sender/receivers are often close to each other, and then MIN works better.
-    nrf = NRF24(pi, ce=25, payload_size=9, channel=100, data_rate=RF24_DATA_RATE.RATE_250KBPS, pa_level=RF24_PA.MIN)
+    nrf = NRF24(pi, ce=25, payload_size=RF24_PAYLOAD.DYNAMIC, channel=100, data_rate=RF24_DATA_RATE.RATE_250KBPS, pa_level=RF24_PA.LOW)
     nrf.set_address_bytes(len(address))
     nrf.open_writing_pipe(address)
-
+    
     # Display the content of NRF24L01 device registers.
     nrf.show_registers()
 
-    count = 0
-    print(f'Send to {address}')
     try:
+        print(f'Send to {address}')
+        count = 0
         while True:
 
             # Emulate that we read temperature and humidity from a sensor, for example
@@ -68,28 +87,15 @@ if __name__ == "__main__":
             # Send the payload to the address specified above.
             nrf.reset_packages_lost()
             nrf.send(payload)
-            try:
-                nrf.wait_until_sent()
-                
-            except TimeoutError:
-                print("Timeout waiting for transmission to complete.")
-                time.sleep(10)
-                continue
-            
-            if nrf.get_packages_lost() == 0:
-                print(f"Success: lost={nrf.get_packages_lost()}, retries={nrf.get_retries()}")
-            else:
-                print(f"Error: lost={nrf.get_packages_lost()}, retries={nrf.get_retries()}")
 
-            # Wait 10 seconds before sending the next reading.
-            time.sleep(10)
+            # When the send is complete (successful or failed), the NRF24L01+ module will trigger an
+            # interrupt by pulling the INT pin low.  We detect that with the callback on the falling
+            # endge as configured above.
+            
+            # Wait 5 seconds before sending the next reading.
+            time.sleep(5)
 
     except:
         traceback.print_exc()
         nrf.power_down()
         pi.stop()
-
-
-
-
-